Creating ProfileViewModelMapper
