+++ EsmorgaiOS/App/Data/UseCase/User/LogOutUserUseCase.swift
@@ -8,20 +8,20 @@
 import Foundation
 
 enum LogOutUserError: Error {
-   case logOutFailed
+    case logOutFailed
 }
 
 typealias LogoutUserResult = Result<Void, LogOutUserError>
 typealias LogoutUserUseCaseAlias = BaseUseCase<Void, LogoutUserResult>
 
 class LogoutUserUseCase: LogoutUserUseCaseAlias {
-
+    
     private var userRepository: UserRepositoryProtocol
-
+    
     init(userRepository: UserRepositoryProtocol = UserRepository()) {
         self.userRepository = userRepository
     }
-
+    
     override func job() async -> LogoutUserResult {
         let result = await userRepository.logoutUser()
         return result ? .success(()) : .failure(.logOutFailed) //Que tipo de error puedo generar aquí?
